Add recommendations logic to user model
This looks at the user's friends attendances and works out if there
are any events that the user does not already know about. If there
are, the code looks for similar artists to the event headliner and
if there is an overlap between these and the user's tracked artists
it creates a recommendation.